move download code into download_data, now controlled by ramp test env variable; updated notebook, expanded data fetch instructions
