+++ tests/test_hash.py
@@ -1,5 +1,5 @@
 import unittest
-import hash_check
+from download_data import check_hash_correct, get_sha256
 import os
 
 class TestHash(unittest.TestCase):
@@ -7,13 +7,13 @@ class TestHash(unittest.TestCase):
         test_dir = os.path.dirname(__file__)
         file = os.path.join(test_dir, 'hash_sample/hash.txt')
         expected_hash = 'a119e9fd61ee902dbd10c8058a7474ea50cf7ae7edfd17fbf400871b0b31e628'
-        self.assertEqual(hash_check.get_sha256(file), expected_hash)
+        self.assertEqual(get_sha256(file), expected_hash)
         return
 
     def test_checkhash(self):
         test_dir = os.path.dirname(__file__)
         file = os.path.join(test_dir, 'hash_sample/hash.txt')
         expected_hash = 'a119e9fd61ee902dbd10c8058a7474ea50cf7ae7edfd17fbf400871b0b31e628'
-        self.assertTrue(hash_check.check_hash_correct(file, expected_hash))
-        self.assertFalse(hash_check.check_hash_correct(file, 'chicken'))
+        self.assertTrue(check_hash_correct(file, expected_hash))
+        self.assertFalse(check_hash_correct(file, 'chicken'))
         return
